[FIX] product_pack: Take into account both pricelist and pack line discount. (#197)
Formerly was only taking into account the product packs discount, now take
into account both discounts.

Also we simplify the way that get_sale_order_line_vals works and now is
simplier and use Odoo's onchange methods to compute the lines values.

[FIX] product_pack: Pack component now have the tax of the company (#198)

Before this fix the product's packs new sale order line have multiple taxes:
one por each company. Actually only need to add the taxes realated to the
sale order's company.

[FIX] product_pack: compute proper components prices (#199)

In some cases the values were prefetched and it was throwing bad prices.
we add prefetch_fields to the context of the packs and that solve it.

[FIX/IMP] product_pack: Allow modify pack and do not reset after confirm website sale order (#201)

* [FIX] product_pack: Do not reset packs components when confirm sale order.

If a sale order is updated from the website whrn confirm the order we do
not expand the pack, we usit as it is.

* [ADD] product_pack: Allow modify pack from backend/website

Now we can choose if the product can be modified in the only int he backend or
if can be modified also in the frontend by the customers.

* [FIX] lint