[FIX] product_pack: Solve the problem with the packs and the user without right to write in products.
This solve the problem when with a user without rights to write in product, you want to search or access to a product variant who are a pack regardless the type or the composition of that one.

We undone this PR https://github.com/OCA/product-pack/pull/42 because not solve completly the problem, with this change now all problems are solved.