[FIX] stock_product_pack: Remove unnecessary and wrong action from tests
The `stock.scheduler.compute` wizard action is run in a different Thread that the main one.
In a test environment, that is problematic, as the test thread ends without actually getting the result from the 2nd that is launched.
This can lead to unclosed cursors and an inconsistent state in the environment.
Besides, for the same reason, we can simply delete the step from the test, as it was not doing anything.

TT32016