[IMP] stock_product_pack: incoming/outgoing stock
Before this commit, the module do not compute the incoming and outgoing
pack stock.
Now is done following this logic:

Stock_on_hand = The number of packs that can be assembled with the stock
components on hand.
Stock_on_hand_inc = The number of packs that can be assembled with the
stock components on hand plus the incoming stock.
Stock_on_hand_out = The number of packs that can be assembled with the
stock components on hand minus the outgoing stock.

Then:

Stock incoming = Stock_on_hand_inc - Stock_on_hand
This is the number of extra packs that we will be able to assemble if
the incoming stock of the components arrive.

Stock outgoing = Stock_on_hand + Stock_on_hand_out
This is the number of packs that we will NOT be able to assemble if
the outgoing stock of the components is delivered

Stock virtual available = available + incoming - outgoing
Note that this is an improvement because now we are following the same
logic for virtual stock as Odoo in normal products.