[FIX] product_pack: compute price pack detailed totalized
For sale orders where has a Pack (detailed and totalized in main product) and the pricelist has an operation (i.e. list_price + 10%), the result was wrong, adding +10% twice.
In this PR, we only calculate the price with pricelist for the components of the pack and not for the pack