[REF] product_pack: Price computation refactor
This commit improves the way the price of the packs is computed to
ensure it aligns with the behavior of normal products in Odoo.

Key improvements include:

1. Refactor Pack Price Computation:

Removed the inheritance from price_compute in favor of a clearer and
more comprehensible _get_product_price method within the Pricelist class

2. 'get_price' -> '_get_pack_line_price':

The get_price method on pack lines has been improved and renamed to
_get_pack_line_price. It is now capable of returning the price of the
 pack line for a specified pricelist.

3. List Price Calculation:

Enhanced the _compute_product_lst_price method to return the list price
of a pack. It now leverages a new method analogous to price_compute,
named _pack_line_price_compute, which can determine the prices of pack
components, including nested packs.